Tier 0: dispatch syevx over algorithm families, add Direct path
`syevx` was a single batched LOBPCG serving every case. For dense input at
moderate neigs/n that is slower than simply running syev and selecting.

- Add SyevxAlgorithm {Auto, Direct, DirectSubset, Filtered, LOBPCG} and
  SyevxParams::method, with a BATCHLAS_SYEVX_ALGORITHM env override.
- Add syevx_select_algorithm: sparse -> LOBPCG; dense n <= 64 or
  neigs/n > 2% -> Direct; otherwise LOBPCG. Unimplemented tiers
  (DirectSubset, Filtered) degrade to their nearest implemented neighbour.
- Add syevx_direct: syev on a private copy of A plus a selection kernel.
  A is left unmodified; ordering matches the LOBPCG path.
- Rename the LOBPCG implementation to syevx_lobpcg (syevx.cc ->
  syevx_lobpcg.cc); syevx.cc is now the dispatcher.

Also fixes a pre-existing workspace bug in the LOBPCG path: the projected
StAS solve runs at Nvecs = 2*block_vectors on the restart iteration, but
only block_vectors and 3*block_vectors were sized. Since the SYEV provider
is chosen per size, the largest matrix does not necessarily need the
largest workspace, so this threw "syev: insufficient workspace for chosen
provider" for e.g. n=60, neigs=3, extra_directions=10. Both the runtime and
buffer-size paths now cover all three sizes. This is what made
SyevxOperationsTest.RandomMatrix fail on main.

Tests: syevx_tests passes under Auto, Direct and LOBPCG (7 tests). Adds
SyevxDirectTest covering eigenvector output and both orderings against a
reference syev. Fixes ComplexShiftInverToeplitzEigenpairs, which sized its
workspace from a CSR view but solved a dense matrix.

Benchmarks: adds BM_SYEVX_Crossover sweeping (n, batch, neigs/n, algorithm)
at a fixed accuracy target, to replace the flop-count thresholds with
measured ones.
